Fix the Python lint gate the build-test lane failed on
The support module raised PublishReportShapeError from five functions
whose docstrings named no Raises section, and built each message inside
the raise expression. Both are lint errors here, and together they made
the build-test lane red on the ubicloud runner while every other check
passed.

Document the five Raises sections, saying what shape each function
refuses rather than restating the exception's name, and bind each
message before raising it.

Splitting the messages out pushed the module past the 400-line limit
Pylint enforces, so the three lading pin readers move to
tests/workflow_contracts/lading_pins.py. They are a coherent unit: the
places the pin is written, read in one module, separate from the
publish-report vocabulary the other two contracts share.
